refactor, shift support logic to _util
